requirements1.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community
pytesseract
senetence_transformers
pdf2image
langdetect